+++ src/game/Object/Player.cpp
@@ -3812,133 +3812,6 @@ void Player::DeleteOldCharacters(uint32 keepDays)
 }
 
 /**
- * @brief Forces or clears rooted movement for the player.
- *
- * @param enable True to root the player; false to unroot them.
- */
-void Player::SetRoot(bool enable)
-{
-    WorldPacket data(enable ? SMSG_FORCE_MOVE_ROOT : SMSG_FORCE_MOVE_UNROOT, GetPackGUID().size() + 4);
-    data << GetPackGUID();
-    data << uint32(0);
-    SendMessageToSet(&data, true);
-}
-
-/**
- * @brief Enables or disables water walking for the player.
- *
- * @param enable True to enable water walking; false to restore normal movement.
- */
-void Player::SetWaterWalk(bool enable)
-{
-    WorldPacket data(enable ? SMSG_MOVE_WATER_WALK : SMSG_MOVE_LAND_WALK, GetPackGUID().size() + 4);
-    data << GetPackGUID();
-    data << uint32(0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Placeholder for levitation support on this client version.
- *
- * @param enable Unused levitation state flag.
- */
-void Player::SetLevitate(bool enable)
-{
-    // TODO: check if there is something similar for 2.4.3.
-    // WorldPacket data;
-    // if (enable)
-    //    data.Initialize(SMSG_MOVE_GRAVITY_DISABLE, 12);
-    // else
-    //    data.Initialize(SMSG_MOVE_GRAVITY_ENABLE, 12);
-    //
-    // data << GetPackGUID();
-    // data << uint32(0);                                      // unk
-    // SendMessageToSet(&data, true);
-
-    // data.Initialize(MSG_MOVE_GRAVITY_CHNG, 64);
-    // data << GetPackGUID();
-    // m_movementInfo.Write(data);
-    // SendMessageToSet(&data, false);
-}
-
-/**
- * @brief Enables or disables flying movement flags for the player.
- *
- * @param enable True to enable flight-related movement flags; false to clear them.
- */
-void Player::SetCanFly(bool enable)
-{
-    WorldPacket data;
-    if (enable)
-    {
-        data.Initialize(SMSG_MOVE_SET_CAN_FLY, 12);
-    }
-    else
-    {
-        data.Initialize(SMSG_MOVE_UNSET_CAN_FLY, 12);
-    }
-
-    data << GetPackGUID();
-    data << uint32(0);                                      // unk
-    SendMessageToSet(&data, true);
-
-    data.Initialize(MSG_MOVE_UPDATE_CAN_FLY, 64);
-    data << GetPackGUID();
-    m_movementInfo.Write(data);
-    SendMessageToSet(&data, false);
-}
-
-/**
- * @brief Enables or disables feather fall movement for the player.
- *
- * @param enable True to enable feather fall; false to restore normal falling.
- */
-void Player::SetFeatherFall(bool enable)
-{
-    WorldPacket data;
-    if (enable)
-    {
-        data.Initialize(SMSG_MOVE_FEATHER_FALL, 8 + 4);
-    }
-    else
-    {
-        data.Initialize(SMSG_MOVE_NORMAL_FALL, 8 + 4);
-    }
-
-    data << GetPackGUID();
-    data << uint32(0);
-    SendMessageToSet(&data, true);
-
-    // start fall from current height
-    if (!enable)
-    {
-        SetFallInformation(0, GetPositionZ());
-    }
-}
-
-/**
- * @brief Enables or disables hover movement for the player.
- *
- * @param enable True to enable hovering; false to disable it.
- */
-void Player::SetHover(bool enable)
-{
-    WorldPacket data;
-    if (enable)
-    {
-        data.Initialize(SMSG_MOVE_SET_HOVER, 8 + 4);
-    }
-    else
-    {
-        data.Initialize(SMSG_MOVE_UNSET_HOVER, 8 + 4);
-    }
-
-    data << GetPackGUID();
-    data << uint32(0);
-    SendMessageToSet(&data, true);
-}
-
-/**
  * @brief Attempts to improve defense skill and refresh defense-derived bonuses.
  */
 void Player::UpdateDefense()
@@ -4725,128 +4598,6 @@ void Player::ModifyArenaPoints(int32 value)
     }
 
     SetArenaPoints(newValue);
-}
-
-/**
- * @brief Loads a player's guild identifier from the database.
- *
- * @param guid The player GUID to query.
- * @return The guild identifier, or zero if none is found.
- */
-uint32 Player::GetGuildIdFromDB(ObjectGuid guid)
-{
-    uint32 lowguid = guid.GetCounter();
-
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `guildid` FROM `guild_member` WHERE `guid`='%u'", lowguid);
-    if (!result)
-    {
-        return 0;
-    }
-
-    uint32 id = result->Fetch()[0].GetUInt32();
-    delete result;
-    return id;
-}
-
-/**
- * @brief Loads a player's guild rank from the database.
- *
- * @param guid The player GUID to query.
- * @return The guild rank, or zero if none is found.
- */
-uint32 Player::GetRankFromDB(ObjectGuid guid)
-{
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `rank` FROM `guild_member` WHERE `guid`='%u'", guid.GetCounter());
-    if (result)
-    {
-        uint32 v = result->Fetch()[0].GetUInt32();
-        delete result;
-        return v;
-    }
-    else
-    {
-        return 0;
-    }
-}
-
-uint32 Player::GetArenaTeamIdFromDB(ObjectGuid guid, ArenaType type)
-{
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `arena_team_member`.`arenateamid` FROM `arena_team_member` JOIN `arena_team` ON `arena_team_member`.`arenateamid` = `arena_team`.`arenateamid` WHERE `guid`='%u' AND `type`='%u' LIMIT 1", guid.GetCounter(), type);
-    if (!result)
-    {
-        return 0;
-    }
-
-    uint32 id = (*result)[0].GetUInt32();
-    delete result;
-    return id;
-}
-
-/**
- * @brief Loads or derives a player's saved zone identifier from the database.
- *
- * @param guid The player GUID to query.
- * @return The resolved zone identifier, or zero on failure.
- */
-uint32 Player::GetZoneIdFromDB(ObjectGuid guid)
-{
-    uint32 lowguid = guid.GetCounter();
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `zone` FROM `characters` WHERE `guid`='%u'", lowguid);
-    if (!result)
-    {
-        return 0;
-    }
-    Field* fields = result->Fetch();
-    uint32 zone = fields[0].GetUInt32();
-    delete result;
-
-    if (!zone)
-    {
-        // stored zone is zero, use generic and slow zone detection
-        result = CharacterDatabase.PQuery("SELECT `map`,`position_x`,`position_y`,`position_z` FROM `characters` WHERE `guid`='%u'", lowguid);
-        if (!result)
-        {
-            return 0;
-        }
-        fields = result->Fetch();
-        uint32 map = fields[0].GetUInt32();
-        float posx = fields[1].GetFloat();
-        float posy = fields[2].GetFloat();
-        float posz = fields[3].GetFloat();
-        delete result;
-
-        zone = sTerrainMgr.GetZoneId(map, posx, posy, posz);
-
-        if (zone > 0)
-        {
-            CharacterDatabase.PExecute("UPDATE `characters` SET `zone`='%u' WHERE `guid`='%u'", zone, lowguid);
-        }
-    }
-
-    return zone;
-}
-
-/**
- * @brief Loads a player's level from the database.
- *
- * @param guid The player GUID to query.
- * @return The stored level, or zero on failure.
- */
-uint32 Player::GetLevelFromDB(ObjectGuid guid)
-{
-    uint32 lowguid = guid.GetCounter();
-
-    QueryResult* result = CharacterDatabase.PQuery("SELECT `level` FROM `characters` WHERE `guid`='%u'", lowguid);
-    if (!result)
-    {
-        return 0;
-    }
-
-    Field* fields = result->Fetch();
-    uint32 level = fields[0].GetUInt32();
-    delete result;
-
-    return level;
 }
 
 /**
